+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/entity/vo/consumer/ConsumerCreateRequestVO.java
@@ -26,6 +26,8 @@ public class ConsumerCreateRequestVO {
   private String orgId;
   private String orgName;
   private String ownerName;
+  private boolean rateLimitEnabled;
+  private int rateLimit;
 
   public String getAppId() {
     return appId;
@@ -73,6 +75,22 @@ public class ConsumerCreateRequestVO {
 
   public void setOwnerName(String ownerName) {
     this.ownerName = ownerName;
+  }
+
+  public boolean isRateLimitEnabled() {
+    return rateLimitEnabled;
+  }
+
+  public void setRateLimitEnabled(boolean rateLimitEnabled) {
+    this.rateLimitEnabled = rateLimitEnabled;
+  }
+
+  public int getRateLimit() {
+    return rateLimit;
+  }
+
+  public void setRateLimit(int rateLimit) {
+    this.rateLimit = rateLimit;
   }
 
 }