+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/openapi/filter/ConsumerAuthenticationFilterTest.java
@@ -16,9 +16,15 @@
  */
 package com.ctrip.framework.apollo.openapi.filter;
 
+import com.ctrip.framework.apollo.openapi.entity.ConsumerToken;
 import com.ctrip.framework.apollo.openapi.util.ConsumerAuditUtil;
 import com.ctrip.framework.apollo.openapi.util.ConsumerAuthUtil;
 
+import java.io.IOException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import javax.servlet.ServletException;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -33,6 +39,9 @@ import org.springframework.http.HttpHeaders;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.atLeast;
+import static org.mockito.Mockito.atLeastOnce;
+import static org.mockito.Mockito.atMost;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -43,11 +52,15 @@ import static org.mockito.Mockito.when;
  */
 @RunWith(MockitoJUnitRunner.class)
 public class ConsumerAuthenticationFilterTest {
+
+  private static final int TOO_MANY_REQUESTS = 429;
+
   private ConsumerAuthenticationFilter authenticationFilter;
   @Mock
   private ConsumerAuthUtil consumerAuthUtil;
   @Mock
   private ConsumerAuditUtil consumerAuditUtil;
+
   @Mock
   private HttpServletRequest request;
   @Mock
@@ -65,8 +78,11 @@ public class ConsumerAuthenticationFilterTest {
     String someToken = "someToken";
     Long someConsumerId = 1L;
 
+    ConsumerToken someConsumerToken = new ConsumerToken();
+    someConsumerToken.setConsumerId(someConsumerId);
+
     when(request.getHeader(HttpHeaders.AUTHORIZATION)).thenReturn(someToken);
-    when(consumerAuthUtil.getConsumerId(someToken)).thenReturn(someConsumerId);
+    when(consumerAuthUtil.getConsumerToken(someToken)).thenReturn(someConsumerToken);
 
     authenticationFilter.doFilter(request, response, filterChain);
 
@@ -80,7 +96,7 @@ public class ConsumerAuthenticationFilterTest {
     String someInvalidToken = "someInvalidToken";
 
     when(request.getHeader(HttpHeaders.AUTHORIZATION)).thenReturn(someInvalidToken);
-    when(consumerAuthUtil.getConsumerId(someInvalidToken)).thenReturn(null);
+    when(consumerAuthUtil.getConsumerToken(someInvalidToken)).thenReturn(null);
 
     authenticationFilter.doFilter(request, response, filterChain);
 
@@ -89,4 +105,102 @@ public class ConsumerAuthenticationFilterTest {
     verify(consumerAuditUtil, never()).audit(eq(request), anyLong());
     verify(filterChain, never()).doFilter(request, response);
   }
+
+
+  @Test
+  public void testRateLimitSuccessfully() throws Exception {
+    String someToken = "some-ratelimit-success-token";
+    Long someConsumerId = 1L;
+    int qps = 5;
+    int durationInSeconds = 3;
+
+    setupRateLimitMocks(someToken, someConsumerId, qps);
+
+    Runnable task = () -> {
+      try {
+        authenticationFilter.doFilter(request, response, filterChain);
+      } catch (IOException e) {
+        throw new RuntimeException(e);
+      } catch (ServletException e) {
+        throw new RuntimeException(e);
+      }
+    };
+
+    int realQps = qps - 1;
+    executeWithQps(realQps, task, durationInSeconds);
+
+    int total = realQps * durationInSeconds;
+
+    verify(consumerAuthUtil, times(total)).storeConsumerId(request, someConsumerId);
+    verify(consumerAuditUtil, times(total)).audit(request, someConsumerId);
+    verify(filterChain, times(total)).doFilter(request, response);
+
+  }
+
+
+  @Test
+  public void testRateLimitPartFailure() throws Exception {
+     String someToken = "some-ratelimit-fail-token";
+    Long someConsumerId = 1L;
+    int qps = 5;
+    int durationInSeconds = 3;
+
+    setupRateLimitMocks(someToken, someConsumerId, qps);
+
+    Runnable task = () -> {
+      try {
+        authenticationFilter.doFilter(request, response, filterChain);
+      } catch (IOException e) {
+        throw new RuntimeException(e);
+      } catch (ServletException e) {
+        throw new RuntimeException(e);
+      }
+    };
+
+    int realQps = qps + 3;
+    executeWithQps(realQps, task, durationInSeconds);
+
+    int leastTimes = qps * durationInSeconds;
+    int mostTimes = realQps * durationInSeconds;
+
+    verify(response, atLeastOnce()).sendError(eq(TOO_MANY_REQUESTS), anyString());
+
+    verify(consumerAuthUtil, atLeast(leastTimes)).storeConsumerId(request, someConsumerId);
+    verify(consumerAuthUtil, atMost(mostTimes)).storeConsumerId(request, someConsumerId);
+    verify(consumerAuditUtil, atLeast(leastTimes)).audit(request, someConsumerId);
+    verify(consumerAuditUtil, atMost(mostTimes)).audit(request, someConsumerId);
+    verify(filterChain, atLeast(leastTimes)).doFilter(request, response);
+    verify(filterChain, atMost(mostTimes)).doFilter(request, response);
+
+  }
+
+
+  private void setupRateLimitMocks(String someToken, Long someConsumerId, int qps) {
+    ConsumerToken someConsumerToken = new ConsumerToken();
+    someConsumerToken.setConsumerId(someConsumerId);
+    someConsumerToken.setRateLimit(qps);
+    someConsumerToken.setToken(someToken);
+
+    when(request.getHeader(HttpHeaders.AUTHORIZATION)).thenReturn(someToken);
+    when(consumerAuthUtil.getConsumerToken(someToken)).thenReturn(someConsumerToken);
+  }
+
+
+  public static void executeWithQps(int qps, Runnable task, int durationInSeconds) {
+    ExecutorService executor = Executors.newFixedThreadPool(qps);
+    long totalTasks = qps * durationInSeconds;
+
+    for (int i = 0; i < totalTasks; i++) {
+      executor.submit(task);
+      try {
+        TimeUnit.MILLISECONDS.sleep(1000 / qps); // Control QPS
+      } catch (InterruptedException e) {
+        Thread.currentThread().interrupt();
+        break;
+      }
+    }
+
+    executor.shutdown();
+  }
+
 }