+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/portal/controller/ConsumerControllerTest.java
@@ -63,7 +63,7 @@ class ConsumerControllerTest {
 
     Mockito.verify(consumerService, Mockito.times(1)).createConsumer(Mockito.any());
     Mockito.verify(consumerService, Mockito.times(1))
-        .generateAndSaveConsumerToken(Mockito.any(), Mockito.any());
+        .generateAndSaveConsumerToken(Mockito.any(), Mockito.any(), Mockito.any());
     Mockito.verify(consumerService, Mockito.times(0))
         .assignCreateApplicationRoleToConsumer(Mockito.any());
     Mockito.verify(consumerService, Mockito.times(1)).getConsumerInfoByAppId(Mockito.any());
@@ -84,14 +84,14 @@ class ConsumerControllerTest {
     {
       ConsumerToken ConsumerToken = new ConsumerToken();
       ConsumerToken.setToken(token);
-      Mockito.when(consumerService.generateAndSaveConsumerToken(Mockito.any(), Mockito.any()))
+      Mockito.when(consumerService.generateAndSaveConsumerToken(Mockito.any(), Mockito.any(), Mockito.any()))
           .thenReturn(ConsumerToken);
     }
     consumerController.create(requestVO, null);
 
     Mockito.verify(consumerService, Mockito.times(1)).createConsumer(Mockito.any());
     Mockito.verify(consumerService, Mockito.times(1))
-        .generateAndSaveConsumerToken(Mockito.any(), Mockito.any());
+        .generateAndSaveConsumerToken(Mockito.any(), Mockito.any(), Mockito.any());
     Mockito.verify(consumerService, Mockito.times(1))
         .assignCreateApplicationRoleToConsumer(Mockito.eq(token));
     Mockito.verify(consumerService, Mockito.times(1)).getConsumerInfoByAppId(Mockito.any());